+++ lib/serega/plugins/metadata/metadata.rb
@@ -46,13 +46,13 @@ class Serega
       # Checks requirements and loads additional plugins
       #
       # @param serializer_class [Class<Serega>] Current serializer class
-      # @param _opts [Hash] loaded plugins opts
+      # @param _opts [Hash] Plugin options
       #
       # @return [void]
       #
       def self.before_load_plugin(serializer_class, **_opts)
         unless serializer_class.plugin_used?(:root)
-          raise SeregaError, "Please load :root plugin first so we can wrap serialization response into top-level hash to add metadata there"
+          raise SeregaError, "Plugin #{plugin_name.inspect} must be loaded after the :root plugin. Please load the :root plugin first"
         end
       end
 
@@ -60,7 +60,7 @@ class Serega
       # Applies plugin code to specific serializer
       #
       # @param serializer_class [Class<Serega>] Current serializer class
-      # @param _opts [Hash] Loaded plugins options
+      # @param _opts [Hash] Plugin options
       #
       # @return [void]
       #
@@ -87,7 +87,7 @@ class Serega
       # Adds config options and runs other callbacks after plugin was loaded
       #
       # @param serializer_class [Class<Serega>] Current serializer class
-      # @param _opts [Hash] loaded plugins opts
+      # @param _opts [Hash] Plugin options
       #
       # @return [void]
       #